Make `Tests` env. type reusable
Resolves #729
