+++ base/src/test/java/io/spine/environment/given/AppEngine.java
@@ -24,37 +24,14 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.base.given;
+package io.spine.environment.given;
+
+import io.spine.environment.CustomEnvironmentType;
 
 /**
- * Determines whether the system is running under Google App Engine Standard environment.
+ * An environment that denotes that the system is running under Google App Engine.
+ *
+ * <p>Leaves the implementation of {@link #enabled()} to subclasses.
  */
-@SuppressWarnings("AccessOfSystemProperties")
-public class AppEngineStandard extends AppEngine {
-
-    private static final String ENV_KEY = "io.spine.base.test.is_appengine";
-
-    @Override
-    protected boolean enabled() {
-        var propertyValue = System.getProperty(ENV_KEY);
-        return activeValue().equalsIgnoreCase(propertyValue);
-    }
-
-    /**
-     * Enables the App Engine Standard environment.
-     */
-    public static void enable() {
-        System.setProperty(ENV_KEY, activeValue());
-    }
-
-    /**
-     * Disables teh App Engine Standard environment.
-     */
-    public static void clear() {
-        System.clearProperty(ENV_KEY);
-    }
-
-    private static String activeValue() {
-        return String.valueOf(true);
-    }
+public abstract class AppEngine extends CustomEnvironmentType {
 }
